Move aux to aux_ (windows checkout problems)
